Implemented AMR cleanup which sprung from GPU branch. Still has linking error.
