+++ vlasovsolver/cpu_1d_ppm_nonuniform.hpp
@@ -29,13 +29,14 @@
 #include "cmath"
 #include "cpu_slope_limiters.hpp"
 #include "cpu_face_estimates.hpp"
+#include "../definitions.h"
 
 using namespace std;
 
 /*
   Compute parabolic reconstruction with an explicit scheme
 */
-inline void compute_ppm_coeff_nonuniform(const Vec * const dv, const Vec * const values, face_estimate_order order, uint k, Vec a[3], const Realv threshold){
+inline void compute_ppm_coeff_nonuniform(const Realf * const dv, const Vec * const values, face_estimate_order order, uint k, Vec a[3], const Realv threshold){
    Vec fv_l; /*left face value*/
    Vec fv_r; /*right face value*/
    compute_filtered_face_values_nonuniform(dv, values, k, order, fv_l, fv_r, threshold);
